samples/python/requirements.txt
opentelemetry-sdk
azure-core-tracing-opentelemetry
azure-monitor-opentelemetry
azure-ai-inference[opentelemetry]
opentelemetry-exporter-otlp
python-dotenv
azure-identity 
azure-ai-projects
azure-ai-inference
openai
semantic-kernel 
opentelemetry-exporter-otlp-proto-grpc
azure-monitor-opentelemetry-exporter
azure-search
azure-search-documents
openai
opentelemetry-instrumentation-openai-v2
opentelemetry-instrumentation-openai